fix: renamed module for compat purposes
Was not working in a Python 2.7 env.
